+++ installer/test/phx_new_umbrella_test.exs
@@ -150,8 +150,17 @@ defmodule Mix.Tasks.Phx.New.UmbrellaTest do
       )
 
       assert_file(web_path(@app, "test/#{@app}_web/controllers/page_controller_test.exs"))
-      assert_file(web_path(@app, "test/#{@app}_web/controllers/error_html_test.exs"))
-      assert_file(web_path(@app, "test/#{@app}_web/controllers/error_json_test.exs"))
+
+      assert_file(
+        web_path(@app, "test/#{@app}_web/controllers/error_html_test.exs"),
+        "async: true"
+      )
+
+      assert_file(
+        web_path(@app, "test/#{@app}_web/controllers/error_json_test.exs"),
+        "async: true"
+      )
+
       assert_file(web_path(@app, "test/support/conn_case.ex"))
       assert_file(web_path(@app, "test/test_helper.exs"))
 
@@ -185,9 +194,10 @@ defmodule Mix.Tasks.Phx.New.UmbrellaTest do
       end)
 
       # assets
-      assert_file(web_path(@app, ".gitignore"), "/priv/static/assets/")
-      assert_file(web_path(@app, ".gitignore"), "#{@app}_web-*.tar")
-      assert_file(web_path(@app, ".gitignore"), ~r/\n$/)
+      assert_file(web_path(@app, ".gitignore"), fn file ->
+        assert file =~ "/priv/static/assets/"
+        assert file =~ "#{@app}_web-*.tar"
+      end)
 
       assert_file(web_path(@app, "assets/css/app.css"), fn file ->
         assert file =~ "lib/phx_umb_web"
@@ -344,7 +354,6 @@ defmodule Mix.Tasks.Phx.New.UmbrellaTest do
 
       # No assets
       assert_file(web_path(@app, ".gitignore"), fn file ->
-        assert file =~ ~r/\n$/
         refute file =~ "/priv/static/assets/"
       end)
 
@@ -507,8 +516,10 @@ defmodule Mix.Tasks.Phx.New.UmbrellaTest do
     in_tmp("new with no_assets", fn ->
       Mix.Tasks.Phx.New.run([@app, "--umbrella", "--no-assets"])
 
-      refute File.read!(web_path(@app, ".gitignore")) |> String.contains?("/priv/static/assets/")
-      assert_file(web_path(@app, ".gitignore"), ~r/\n$/)
+      assert_file(web_path(@app, ".gitignore"), fn file ->
+        refute file =~ "/priv/static/assets/"
+      end)
+
       assert_file(web_path(@app, "priv/static/assets/js/app.js"))
       assert_file(web_path(@app, "priv/static/assets/css/app.css"))
       assert_file(web_path(@app, "priv/static/favicon.ico"))
@@ -625,6 +636,16 @@ defmodule Mix.Tasks.Phx.New.UmbrellaTest do
       assert_file(root_path(app, "config/runtime.exs"), [~r/url: database_url/])
 
       assert_file(web_path(app, "test/support/conn_case.ex"), "DataCase.setup_sandbox(tags)")
+
+      assert_file(
+        web_path(app, "test/custom_path_web/controllers/error_html_test.exs"),
+        fn file -> refute file =~ "async: true" end
+      )
+
+      assert_file(
+        web_path(app, "test/custom_path_web/controllers/error_json_test.exs"),
+        fn file -> refute file =~ "async: true" end
+      )
     end)
   end
 
@@ -651,6 +672,16 @@ defmodule Mix.Tasks.Phx.New.UmbrellaTest do
       assert_file(root_path(app, "config/runtime.exs"), [~r/database: database_path/])
 
       assert_file(web_path(app, "test/support/conn_case.ex"), "DataCase.setup_sandbox(tags)")
+
+      assert_file(
+        web_path(app, "test/custom_path_web/controllers/error_html_test.exs"),
+        fn file -> refute file =~ "async: true" end
+      )
+
+      assert_file(
+        web_path(app, "test/custom_path_web/controllers/error_json_test.exs"),
+        fn file -> refute file =~ "async: true" end
+      )
 
       assert_file(root_path(app, ".gitignore"), "*.db")
       assert_file(root_path(app, ".gitignore"), "*.db-*")
@@ -679,6 +710,16 @@ defmodule Mix.Tasks.Phx.New.UmbrellaTest do
       assert_file(root_path(app, "config/runtime.exs"), [~r/url: database_url/])
 
       assert_file(web_path(app, "test/support/conn_case.ex"), "DataCase.setup_sandbox(tags)")
+
+      assert_file(
+        web_path(app, "test/custom_path_web/controllers/error_html_test.exs"),
+        fn file -> refute file =~ "async: true" end
+      )
+
+      assert_file(
+        web_path(app, "test/custom_path_web/controllers/error_json_test.exs"),
+        fn file -> refute file =~ "async: true" end
+      )
     end)
   end
 
@@ -822,8 +863,8 @@ defmodule Mix.Tasks.Phx.New.UmbrellaTest do
         assert_file("another/lib/another/endpoint.ex", ~r/defmodule Another.Endpoint do/)
 
         assert_file("another/test/another/controllers/page_controller_test.exs")
-        assert_file("another/test/another/controllers/error_html_test.exs")
-        assert_file("another/test/another/controllers/error_json_test.exs")
+        assert_file("another/test/another/controllers/error_html_test.exs", "async: true")
+        assert_file("another/test/another/controllers/error_json_test.exs", "async: true")
         assert_file("another/test/support/conn_case.ex")
         assert_file("another/test/test_helper.exs")
 
@@ -844,7 +885,7 @@ defmodule Mix.Tasks.Phx.New.UmbrellaTest do
         assert_file("another/lib/another/components/layouts/root.html.heex")
 
         # assets
-        assert_file("another/.gitignore", ~r/\n$/)
+        assert_file("another/.gitignore")
         assert_file("another/priv/static/favicon.ico")
         assert_file("another/assets/css/app.css")