+++ test/phoenix/router/pipeline_test.exs
@@ -174,4 +174,56 @@ defmodule Phoenix.Router.PipelineTest do
       end
     end
   end
+
+  test "pipe_through accepts pipelines and all plug forms" do
+    defmodule PassThroughRouter do
+      use Phoenix.Router
+
+      def push_stack(conn, value) do
+        assign(conn, :stack, [value | conn.assigns[:stack] || []])
+      end
+
+      def function_plug(conn, opts) do
+        PassThroughRouter.push_stack(conn, {:function_plug, opts})
+      end
+
+      defmodule ModulePlug do
+        def init(opts), do: opts
+
+        def call(conn, opts) do
+          PassThroughRouter.push_stack(conn, {:module_plug, opts})
+        end
+      end
+
+      pipeline :pipeline1 do
+        plug :push_stack, :pipeline1
+      end
+
+      scope "/" do
+        pipe_through [
+          :pipeline1,
+          :function_plug,
+          {:function_plug, []},
+          {:function_plug, [:opt1]},
+          ModulePlug,
+          {ModulePlug, []},
+          {ModulePlug, [:opt1]}
+        ]
+
+        get "/hello", SampleController, :index
+      end
+    end
+
+    conn = call(PassThroughRouter, :get, "/hello")
+
+    assert Enum.reverse(conn.assigns[:stack]) == [
+             :pipeline1,
+             {:function_plug, []},
+             {:function_plug, []},
+             {:function_plug, [:opt1]},
+             {:module_plug, []},
+             {:module_plug, []},
+             {:module_plug, [:opt1]}
+           ]
+  end
 end